requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
api4jenkins==2.0.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
multi-key-dict==2.0.3
multidict==6.0.4
pbr==6.0.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
ujenkins==0.10.2
urllib3==2.1.0
yarl==1.9.4
